refactor(encoding): extract RowMatchGenerator into row/ module
Move the `RowMatchGenerator` struct, its inherent impl, and the row-
specific constants out of `match_generator.rs` into a dedicated
`row::` submodule (~390 LoC). Mechanical move — names, fields, and
method bodies are preserved; the visibility on the relocated items is
opened to `pub(crate)` so:

- `MatchGeneratorDriver` in `match_generator` can keep instantiating
  the row backend and accessing the row tables during reset/configure
  paths (`row.row_heads`, `row.row_positions`, `row.row_tags`,
  `row.lazy_depth`, `row.offset_hist`, `row.set_hash_bits`).
- The encoder integration tests can keep inspecting internal state
  (`row.search_depth`, `row.hash_kernel`, `row.row_hash_log`,
  `row.{ensure_tables, live_history, hash_and_row, best_match,
  pick_lazy_match, repcode_candidate, row_candidate}`).

`match_generator.rs` shrinks by 388 lines (9905 → 9462) and gains a
single `use super::row::RowMatchGenerator` import; the row-specific
constants (`ROW_*`, `INCOMPRESSIBLE_SKIP_STEP`, `RowConfig`) and the
`MatchGenerator::common_prefix_len` static method are promoted to
`pub(crate)` so the relocated matcher can keep calling them.

Build, lint (x86_64 + aarch64), and the level22 ratio gate stay green.

Part of #111 Phase 1b.
